refactor: refactor NNS recovery test to support more scenarios
