+++ PULL_REQUEST_BAZEL_TARGETS
@@ -68,7 +68,7 @@ rs/backup/*.rs //rs/tests/consensus/backup/...
 
 rs/recovery/src/*.rs //rs/tests/consensus/subnet_recovery/...
                      //rs/tests/consensus:subnet_splitting_test_colocate
-                     //rs/tests/nested:nns_recovery_test
+                     //rs/tests/nested/nns_recovery_test/...
 
 rs/recovery/subnet_splitting/*.rs //rs/tests/consensus:subnet_splitting_test_colocate
