+++ rs/tests/nested/Cargo.toml
@@ -8,14 +8,12 @@ documentation.workspace = true
 
 [dependencies]
 anyhow = { workspace = true }
-futures = { workspace = true }
 canister-test = { path = "../../rust_canisters/canister_test" }
 ic-canister-client = { path = "../../canister_client" }
 ic-nervous-system-common-test-keys = { path = "../../nervous_system/common/test_keys" }
 ic-nns-common = { path = "../../nns/common" }
 ic-nns-test-utils = { path = "../../nns/test_utils" }
 ic-protobuf = { path = "../../protobuf" }
-ic-recovery = { path = "../../recovery" }
 ic-registry-subnet-type = { path = "../../registry/subnet_type" }
 ic-registry-nns-data-provider = { path = "../../registry/nns_data_provider" }
 ic-registry-keys = { path = "../../registry/keys" }
@@ -23,12 +21,14 @@ ic-system-test-driver = { path = "../driver" }
 ic-types = { path = "../../types/types" }
 ic_consensus_system_test_utils = { path = "../consensus/utils" }
 prost = { workspace = true }
-rand = { workspace = true }
 reqwest = { workspace = true }
 regex = { workspace = true }
-sha2 = { workspace = true }
 slog = { workspace = true }
 url = { workspace = true }
+
+[[lib]]
+name = "nested"
+path = "src/lib.rs"
 
 [[bin]]
 name = "registration"
