+++ rs/tests/nested/nns_recovery/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "nested-nns-recovery"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+description.workspace = true
+documentation.workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+futures = { workspace = true }
+ic_consensus_system_test_utils = { path = "../../consensus/utils" }
+ic-recovery = { path = "../../../recovery" }
+ic-system-test-driver = { path = "../../../driver" }
+ic-types = { path = "../../../types/types" }
+nested = { path = "../nested" }
+rand = { workspace = true }
+sha2 = { workspace = true }
+slog = { workspace = true }
+
+[lib]
+name = "ic_nested_nns_recovery_common"
+path = "common.rs"
+
+[[bin]]
+name = "ic-systest-nested-nns-recovery-remote"
+path = "nr_remote.rs"
