+++ rs/tests/nested/registration.rs
@@ -4,7 +4,7 @@ use std::time::Duration;
 
 fn main() -> Result<()> {
     SystemTestGroup::new()
-        .with_setup(|env| nested::config(env, 1, None))
+        .with_setup(nested::config)
         .add_test(systest!(nested::registration))
         .with_timeout_per_test(Duration::from_secs(20 * 60))
         .with_overall_timeout(Duration::from_secs(30 * 60))
